requirements.txt
numpy>=1.24
scipy>=1.10
opencv-python>=4.8
pydantic>=2.5
PyYAML>=6.0
pytest>=7.4
torch>=2.1
ultralytics>=8.2
fastapi>=0.110
uvicorn>=0.27
psutil>=5.9
